+++ topological_adam/__init__.py
@@ -1,5 +1,6 @@
 from .analysis import ExperimentConfig, print_comparison_report, run_comparison, run_instrumented_training
 from .optimizer import TopologicalAdam, TopologicalAdamV2
+from .sds import TopologicalAdamSDS
 from .stopping import ReconnectionStoppingRule, StopDecision
 
 __all__ = [
@@ -8,6 +9,7 @@ __all__ = [
     "StopDecision",
     "TopologicalAdam",
     "TopologicalAdamV2",
+    "TopologicalAdamSDS",
     "print_comparison_report",
     "run_comparison",
     "run_instrumented_training",